refcator spreadsheet to use table and roow based approach rather than column based
