+++ web/components/issues/issue-layouts/spreadsheet/columns/assignee-column.tsx
@@ -1,56 +1,31 @@
 import React from "react";
-// hooks
-import { useIssueDetail } from "hooks/store";
 // components
 import { ProjectMemberDropdown } from "components/dropdowns";
 // types
 import { TIssue } from "@plane/types";
 
 type Props = {
-  issueId: string;
+  issue: TIssue;
   onChange: (issue: TIssue, data: Partial<TIssue>) => void;
-  expandedIssues: string[];
   disabled: boolean;
 };
 
-export const SpreadsheetAssigneeColumn: React.FC<Props> = ({ issueId, onChange, expandedIssues, disabled }) => {
-  const isExpanded = expandedIssues.indexOf(issueId) > -1;
-
-  const { subIssues: subIssuesStore, issue } = useIssueDetail();
-
-  const issueDetail = issue.getIssueById(issueId);
-  const subIssues = subIssuesStore.subIssuesByIssueId(issueId);
+export const SpreadsheetAssigneeColumn: React.FC<Props> = (props: Props) => {
+  const { issue, onChange, disabled } = props;
 
   return (
-    <>
-      {issueDetail && (
-        <div className="h-11 border-b-[0.5px] border-custom-border-200">
-          <ProjectMemberDropdown
-            value={issueDetail?.assignee_ids ?? []}
-            onChange={(data) => onChange(issueDetail, { assignee_ids: data })}
-            projectId={issueDetail?.project_id}
-            disabled={disabled}
-            multiple
-            placeholder="Assignees"
-            buttonVariant={issueDetail.assignee_ids.length > 0 ? "transparent-without-text" : "transparent-with-text"}
-            buttonClassName="text-left"
-            buttonContainerClassName="w-full"
-          />
-        </div>
-      )}
-
-      {isExpanded &&
-        subIssues &&
-        subIssues.length > 0 &&
-        subIssues.map((subIssueId) => (
-          <SpreadsheetAssigneeColumn
-            key={subIssueId}
-            issueId={subIssueId}
-            onChange={onChange}
-            expandedIssues={expandedIssues}
-            disabled={disabled}
-          />
-        ))}
-    </>
+    <div className="h-11 border-b-[0.5px] border-custom-border-200">
+      <ProjectMemberDropdown
+        value={issue?.assignee_ids ?? []}
+        onChange={(data) => onChange(issue, { assignee_ids: data })}
+        projectId={issue?.project_id}
+        disabled={disabled}
+        multiple
+        placeholder="Assignees"
+        buttonVariant={issue.assignee_ids.length > 0 ? "transparent-without-text" : "transparent-with-text"}
+        buttonClassName="text-left"
+        buttonContainerClassName="w-full"
+      />
+    </div>
   );
 };
