+++ web/components/issues/issue-layouts/spreadsheet/columns/due-date-column.tsx
@@ -1,6 +1,4 @@
 import React from "react";
-// hooks
-import { useIssueDetail } from "hooks/store";
 // components
 import { DateDropdown } from "components/dropdowns";
 // helpers
@@ -9,49 +7,25 @@ import { renderFormattedPayloadDate } from "helpers/date-time.helper";
 import { TIssue } from "@plane/types";
 
 type Props = {
-  issueId: string;
+  issue: TIssue;
   onChange: (issue: TIssue, data: Partial<TIssue>) => void;
-  expandedIssues: string[];
   disabled: boolean;
 };
 
-export const SpreadsheetDueDateColumn: React.FC<Props> = ({ issueId, onChange, expandedIssues, disabled }) => {
-  const isExpanded = expandedIssues.indexOf(issueId) > -1;
-
-  // const { subIssues, isLoading, mutateSubIssues } = useSubIssue(issue.project_id, issue.id, isExpanded);
-  const { subIssues: subIssuesStore, issue } = useIssueDetail();
-
-  const issueDetail = issue.getIssueById(issueId);
-  const subIssues = subIssuesStore.subIssuesByIssueId(issueId);
+export const SpreadsheetDueDateColumn: React.FC<Props> = (props: Props) => {
+  const { issue, onChange, disabled } = props;
 
   return (
-    <>
-      {issueDetail && (
-        <div className="h-11 border-b-[0.5px] border-custom-border-200">
-          <DateDropdown
-            value={issueDetail.target_date}
-            onChange={(data) => onChange(issueDetail, { target_date: data ? renderFormattedPayloadDate(data) : null })}
-            disabled={disabled}
-            placeholder="Due date"
-            buttonVariant="transparent-with-text"
-            buttonClassName="rounded-none text-left"
-            buttonContainerClassName="w-full"
-          />
-        </div>
-      )}
-
-      {isExpanded &&
-        subIssues &&
-        subIssues.length > 0 &&
-        subIssues.map((subIssueId) => (
-          <SpreadsheetDueDateColumn
-            key={subIssueId}
-            issueId={subIssueId}
-            onChange={onChange}
-            expandedIssues={expandedIssues}
-            disabled={disabled}
-          />
-        ))}
-    </>
+    <div className="h-11 border-b-[0.5px] border-custom-border-200">
+      <DateDropdown
+        value={issue.target_date}
+        onChange={(data) => onChange(issue, { target_date: data ? renderFormattedPayloadDate(data) : null })}
+        disabled={disabled}
+        placeholder="Due date"
+        buttonVariant="transparent-with-text"
+        buttonClassName="rounded-none text-left"
+        buttonContainerClassName="w-full"
+      />
+    </div>
   );
 };
