--- web/components/issues/issue-layouts/spreadsheet/columns/issue/index.ts
@@ -1,2 +1,0 @@
-export * from "./spreadsheet-issue-column";
-export * from "./issue-column";
