+++ web/components/issues/issue-layouts/spreadsheet/index.ts
@@ -1,5 +1,4 @@
 export * from "./columns";
 export * from "./roots";
-export * from "./spreadsheet-column";
 export * from "./spreadsheet-view";
 export * from "./quick-add-issue-form";
